style & refactor: moved PolicyButton to seperate file, and changed text spacing.
